mysite/requirements.txt
altgraph==0.17
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
boto3==1.20.26
botocore==1.23.26
cffi==1.15.0
click==8.0.0
colorama==0.4.4
cryptography==36.0.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.0
django-appconf==1.0.4
django-cors-headers==3.10.1
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-image-cropping==1.5.0
django-rest-knox==4.1.0
django-storages==1.12.3
djangorestframework==3.13.1
filelock==3.0.12
Flask==2.0.0
Flask-MySQLdb==0.2.0
future==0.18.2
gunicorn==20.1.0
itsdangerous==2.0.0
Jinja2==3.0.0
jmespath==0.10.0
MarkupSafe==2.0.0
mutagen==1.45.1
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.21.0
olefile==0.46
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
passlib==1.7.4
pefile==2021.5.24
Pillow==8.3.1
platformdirs==2.2.0
protobuf==3.17.0
psycopg2==2.8.6
pycoin==0.90.20201031
pycparser==2.21
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.2
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.1
pywin32-ctypes==0.2.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
tzdata==2021.5
urllib3==1.26.7
virtualenv==20.7.2
Werkzeug==2.0.0
whitenoise==5.3.0
